Split up classes in utils file
