+++ src/base/mod.rs
@@ -1,6 +1,19 @@
+mod length_prefixed_vec;
 mod local_context;
-pub mod message_utils;
-pub mod raft_utils;
-pub mod utils;
+mod message_utils;
+mod raft_utils;
+mod response_types;
+mod utils;
 
+pub use length_prefixed_vec::LengthPrefixedVec;
 pub use local_context::LocalContext;
+pub use message_utils::{
+    access_type, accessed_inode, distribution_requirement, raft_group, request_locks, AccessType,
+    DistributionRequirement,
+};
+pub use raft_utils::node_contains_raft_group;
+pub use response_types::{FlatBufferResponse, FlatBufferWithResponse, ResultResponse};
+pub use utils::{
+    check_access, empty_response, finalize_request, finalize_request_without_prefix,
+    finalize_response, finalize_response_without_prefix, node_id_from_address, response_or_error,
+};
