+++ src/client/peer_client.rs
@@ -2,9 +2,7 @@ use std::net::SocketAddr;
 
 use flatbuffers::FlatBufferBuilder;
 
-use crate::base::utils::{
-    finalize_request, response_or_error, FlatBufferWithResponse, LengthPrefixedVec,
-};
+use crate::base::{finalize_request, response_or_error, FlatBufferWithResponse, LengthPrefixedVec};
 use crate::generated::*;
 use byteorder::{ByteOrder, LittleEndian};
 use futures::future::{ok, ready, Either};
