+++ src/storage/local/mod.rs
@@ -1,6 +1,7 @@
 mod data_storage;
 mod file_storage;
 mod metadata_storage;
+mod response_helpers;
 
 pub use file_storage::FileStorage;
 pub use metadata_storage::ROOT_INODE;
