+++ src/storage/raft_group_manager.rs
@@ -1,4 +1,4 @@
-use crate::base::raft_utils::node_contains_raft_group;
+use crate::base::node_contains_raft_group;
 use crate::base::LocalContext;
 use crate::storage::raft_node::RaftNode;
 use std::collections::HashMap;
